+++ Backend/src/main/java/org/kmr/backend/review/repository/ReviewRepository.java
@@ -1,5 +1,6 @@
-package org.kmr.backend.review;
+package org.kmr.backend.review.repository;
 
+import org.kmr.backend.review.domain.ReviewEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
